Refactoring

+++ Repository/MatchOddRepository.cs
@@ -5,16 +5,16 @@ using SportBet.Models;
 
 namespace SportBet.Repository
 {
-    public class MatchOddsRepository : IMatchOddsRepository
+    public class MatchOddRepository : IMatchOddRepository
     {
         private readonly MatchContext _dbContext;
 
-        public MatchOddsRepository(MatchContext dbContext)
+        public MatchOddRepository(MatchContext dbContext)
         {
             _dbContext = dbContext;
         }
 
-        public void CreateMatchOdds(MatchOdds matchOdds)
+        public void CreateMatchOdds(MatchOdd matchOdds)
         {
             _dbContext.Add(matchOdds);
             Save();
@@ -23,20 +23,21 @@ namespace SportBet.Repository
         public void DeleteMatchOdds(string id)
         {
             var matchOdds = _dbContext.MatchOdds.Find(id);
-            _dbContext.MatchOdds.Remove(matchOdds);           // handle nulls ???
-            Save();
+            if (matchOdds != null)
+            {
+                _dbContext.MatchOdds.Remove(matchOdds);
+                Save();
+            }
         }
 
-        public IEnumerable<MatchOdds> GetMatchesOdds()
+        public IEnumerable<MatchOdd> GetMatchesOdds()
         {
             return _dbContext.MatchOdds.ToList();
         }
 
-        public MatchOdds GetMatchOdds(string id)
+        public MatchOdd GetMatchOdds(string id)
         {
             return _dbContext.MatchOdds.Find(id);
-
-            // handle nulls ???
         }
 
         public void Save()
@@ -44,7 +45,7 @@ namespace SportBet.Repository
             _dbContext.SaveChanges();
         }
 
-        public void UpdateMatchOdds([FromBody] MatchOdds matchOdds)
+        public void UpdateMatchOdds([FromBody] MatchOdd matchOdds)
         {
             _dbContext.Entry(matchOdds).State = EntityState.Modified;
             Save();
